change match struct to use float instead of double

+++ src/Structs.hpp
@@ -5,8 +5,8 @@
  * the camera and the projector.
  */
 struct Match {
-  double cam_u;  // Camera pixel X
-  double cam_v;  // Camera pixel Y
-  double proj_u; // Projector pixel X
-  double proj_v; // Projector pixel Y
+  float cam_u;  // Camera pixel X
+  float cam_v;  // Camera pixel Y
+  float proj_u; // Projector pixel X
+  float proj_v; // Projector pixel Y
 };
